make manifests

+++ apis/flux/v1alpha1/podinfo_types.go
@@ -28,11 +28,14 @@ type PodInfoSpec struct {
 	// INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
 	// Important: Run "make" to regenerate code after modifying this file
 
-	// Foo is an example field of PodInfo. Edit podinfo_types.go to remove/update
-	PodName    string  `json:"podname,omitempty"`
-	NodeName   string  `json:"nodename,omitempty"`
-	CpuLimit   float32 `json:"cpu_limit,omitempty"`
-	CpuRequest float32 `json:"cpu_request,omitempty"`
+	// The name for pod
+	PodName string `json:"podname,omitempty"`
+	// The assigned nodename of the pod
+	NodeName string `json:"nodename,omitempty"`
+	// cpu limit (roundup)
+	CpuLimit int `json:"cpu_limit,omitempty"`
+	// cpu request (roundup)
+	CpuRequest int `json:"cpu_request,omitempty"`
 }
 
 // PodInfoStatus defines the observed state of PodInfo
